backend/requirements.txt
fastapi
uvicorn
opencv-python
pillow
supabase
python-multipart
numpy
python-dotenv
pydantic-settings
deepface
tf-keras
retina-face
